requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
aiodns==2.0.0
amqp==2.5.2               # via kombu
asgiref==3.2.3            # via django
billiard==3.6.3.0         # via celery
celery[redis]==4.4.1
certifi==2019.11.28       # via requests
cffi==1.14.0              # via pycares
chardet==3.0.4            # via mbstrdecoder, requests
coverage==5.0.3
django-cors-headers==3.2.1
django-environ==0.4.5     # via easyconf
django-redis==4.11.0
django-rest-framework==0.1.0
django-seed==0.2.2
django-test-plus==1.4.0
django==3.0.4
djangorestframework==3.11.0  # via django-rest-framework, drf-nested-routers
drf-nested-routers==0.91
easyconf==1.3.1
faker==4.0.1              # via django-seed
gitdb==4.0.2              # via gitpython
gitpython==3.1.0          # via easyconf
honcho==1.0.1
humanreadable==0.1.0      # via pingparsing
idna==2.9                 # via requests
importlib-metadata==1.5.0  # via kombu
influxdb==5.2.3
kombu==4.6.8              # via celery
mbstrdecoder==0.8.4       # via subprocrunner, typepy
pingparsing==1.0.1
pycares==3.1.1            # via aiodns
pycparser==2.20           # via cffi
git+https://github.com/zodman/pydnsbl.git
pyparsing==2.4.6          # via pingparsing
python-dateutil==2.8.1    # via faker, influxdb
pytz==2019.3              # via celery, django, influxdb
redis==3.4.1              # via celery, django-redis
requests==2.23.0
ruamel.yaml==0.15.100     # via easyconf
six==1.14.0               # via influxdb, python-dateutil, typepy
smmap==3.0.1              # via gitdb
sqlparse==0.3.1           # via django
subprocrunner==1.1.0      # via pingparsing
text-unidecode==1.3       # via faker
typepy==0.6.6             # via humanreadable, pingparsing
urllib3==1.25.8           # via requests
vine==1.3.0               # via amqp, celery
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
